Extract `URL.collect()`
Will be used in more places.
